[13] refactors tree generating implementation into its own file
